requirements.txt
transformers
datasets
nltk
jieba
accelerate
SentencePiece
openai==0.28
matplotlib
Cython
Pillow
sacrebleu
sentence-transformers
networkx
translate
tiktoken
seaborn